add glowing image component
